+++ src/pmp/algorithms/DifferentialGeometry.cpp
@@ -37,6 +37,16 @@ Scalar surface_area(const SurfaceMesh& mesh)
     return A;
 }
 
+Scalar edge_area(const SurfaceMesh& mesh, Edge e)
+{
+    Scalar A(0.0);
+    Face f0 = mesh.face(e, 0);
+    Face f1 = mesh.face(e, 1);
+    if (f0.is_valid()) A += face_area(mesh, f0) / mesh.valence(f0);
+    if (f1.is_valid()) A += face_area(mesh, f1) / mesh.valence(f1);
+    return A;
+}
+
 Scalar voronoi_area(const SurfaceMesh& mesh, Vertex v)
 {
     Scalar A(0.0);